meet_consumer/requirements.txt
python-telegram-bot[job-queue]==21.4
pika==1.3.2